+++ src/main/java/edu/kit/datamanager/metastore2/web/ISchemaRegistryController.java
@@ -30,6 +30,7 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springdoc.core.converters.models.PageableAsQueryParam;
+import org.springframework.boot.actuate.info.InfoContributor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -53,7 +54,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 @ApiResponses(value = {
   @ApiResponse(responseCode = "401", description = "Unauthorized is returned if authorization in required but was not provided."),
   @ApiResponse(responseCode = "403", description = "Forbidden is returned if the caller has no sufficient privileges.")})
-public interface ISchemaRegistryController {
+public interface ISchemaRegistryController extends InfoContributor {
 
   @Operation(summary = "Register a schema document and its record.", description = "This endpoint allows to register a schema document and its record. "
           + "The record must contain at least an unique identifier (schemaId) and the type of the schema (type).",
