+++ src/main/java/edu/kit/datamanager/metastore2/web/impl/MetadataControllerImpl.java
@@ -35,6 +35,7 @@ import edu.kit.datamanager.metastore2.domain.LinkedMetadataRecord;
 import edu.kit.datamanager.metastore2.domain.MetadataRecord;
 import edu.kit.datamanager.metastore2.domain.MetadataSchemaRecord;
 import edu.kit.datamanager.metastore2.domain.ResourceIdentifier;
+import edu.kit.datamanager.metastore2.util.ActuatorUtil;
 import edu.kit.datamanager.metastore2.util.MetadataRecordUtil;
 import edu.kit.datamanager.metastore2.util.MetadataSchemaRecordUtil;
 import edu.kit.datamanager.metastore2.web.IMetadataController;
@@ -58,11 +59,13 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.net.URL;
 import java.nio.file.Path;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
 import javax.servlet.http.HttpServletRequest;
@@ -70,6 +73,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.actuate.info.Info;
 import org.springframework.cloud.gateway.mvc.ProxyExchange;
 import org.springframework.context.annotation.Bean;
 import org.springframework.core.io.FileSystemResource;
@@ -107,11 +111,9 @@ public class MetadataControllerImpl implements IMetadataController {
   final JsonNodeFactory factory = JsonNodeFactory.instance;
 
   private static final Logger LOG = LoggerFactory.getLogger(MetadataControllerImpl.class);
-  @Autowired
-  private ApplicationProperties applicationProperties;
 
   @Autowired
-  private ILinkedMetadataRecordDao metadataRecordDao;
+  private final ILinkedMetadataRecordDao metadataRecordDao;
 
   private final MetastoreConfiguration metadataConfig;
   @Autowired
@@ -470,6 +472,19 @@ public class MetadataControllerImpl implements IMetadataController {
     return new ResponseEntity<>(HttpStatus.NO_CONTENT);
   }
 
+  @Override
+  public void contribute(Info.Builder builder) {
+    LOG.trace("Check for MetadataRepo actuator information...");
+
+    URL basePath = metadataConfig.getBasepath();
+    Map<String, String> details = ActuatorUtil.testDirectory(basePath);
+
+    if (!details.isEmpty()) {
+      details.put("No of metadata documents", Long.toString(MetadataRecordUtil.getNoOfDocuments()));
+      builder.withDetail("metadataRepo", details);
+    }
+  }
+  
   @PostMapping("/{schemaId}/search")
   @Operation(summary = "Search for metadata refered to given schemaId(s) using elastic query language.", description = "Obtaining search results. "
           + "Depending on a user's role, the results are filtered.",
